+++ kusto/ingest/internal/queued/queued_test.go
@@ -10,9 +10,9 @@ import (
 	"testing"
 
 	"github.com/Azure/azure-kusto-go/kusto/data/errors"
+	"github.com/Azure/azure-kusto-go/kusto/ingest/ingestoptions"
 	"github.com/Azure/azure-kusto-go/kusto/ingest/internal/properties"
 	"github.com/Azure/azure-kusto-go/kusto/ingest/internal/utils"
-	"github.com/Azure/azure-kusto-go/kusto/ingest/source"
 
 	"github.com/stretchr/testify/assert"
 
@@ -58,13 +58,13 @@ func TestCompressionDiscovery(t *testing.T) {
 
 	tests := []struct {
 		input string
-		want  source.CompressionType
+		want  ingestoptions.CompressionType
 	}{
-		{"https://somehost.somedomain.com:8080/v1/somestuff/file.gz", source.GZIP},
-		{"https://somehost.somedomain.com:8080/v1/somestuff/file.zip", source.ZIP},
-		{"/path/to/a/file.gz", source.GZIP},
-		{"/path/to/a/file.zip", source.ZIP},
-		{"/path/to/a/file", source.CTNone},
+		{"https://somehost.somedomain.com:8080/v1/somestuff/file.gz", ingestoptions.GZIP},
+		{"https://somehost.somedomain.com:8080/v1/somestuff/file.zip", ingestoptions.ZIP},
+		{"/path/to/a/file.gz", ingestoptions.GZIP},
+		{"/path/to/a/file.zip", ingestoptions.ZIP},
+		{"/path/to/a/file", ingestoptions.CTNone},
 	}
 
 	for _, test := range tests {
@@ -299,6 +299,66 @@ func TestIsLocalPath(t *testing.T) {
 
 			assert.NoError(t, err)
 
+			assert.Equal(t, test.want, got)
+		})
+	}
+}
+
+func TestShouldCompress(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name  string
+		props *properties.All
+		want  bool
+	}{
+		{
+			name: "Some file",
+			props: &properties.All{Source: properties.SourceOptions{CompressionType: ingestoptions.CTUnknown,
+				OriginalSource: "https://somehost.somedomain.com:8080/v1/somestuff/file"}},
+			want: true,
+		},
+		{
+			name: "Some file2",
+			props: &properties.All{Source: properties.SourceOptions{CompressionType: ingestoptions.CTNone,
+				OriginalSource: "https://somehost.somedomain.com:8080/v1/somestuff/file"}},
+			want: true,
+		},
+		{
+			name: "Provided compression type is GZIP",
+			props: &properties.All{Source: properties.SourceOptions{CompressionType: ingestoptions.GZIP,
+				OriginalSource: "https://somehost.somedomain.com:8080/v1/somestuff/file"}},
+			want: false,
+		},
+		{
+			name: "Guess by name is GZIP",
+			props: &properties.All{Source: properties.SourceOptions{CompressionType: ingestoptions.CTUnknown,
+				OriginalSource: "https://somehost.somedomain.com:8080/v1/somestuff/file.gz"}},
+			want: false,
+		},
+		{
+			name: "DontCompress is true",
+			props: &properties.All{Source: properties.SourceOptions{CompressionType: ingestoptions.CTNone,
+				DontCompress:   true,
+				OriginalSource: "https://somehost.somedomain.com:8080/v1/somestuff/file"}},
+			want: false,
+		},
+		{
+			name: "Binary format",
+			props: &properties.All{Source: properties.SourceOptions{CompressionType: ingestoptions.CTNone,
+				OriginalSource: "https://somehost.somedomain.com:8080/v1/somestuff/file.avro"}},
+			want: false,
+		},
+	}
+
+	for _, test := range tests {
+		test := test // capture
+		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
+			CompleteFormatFromFileName(test.props, test.props.Source.OriginalSource)
+
+			got := ShouldCompress(test.props,
+				utils.CompressionDiscovery(test.props.Source.OriginalSource))
 			assert.Equal(t, test.want, got)
 		})
 	}